# .github/fern/openapi/openapi.json
{
  "components": {
...
    "schemas": {
...
      "AgentWriteRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AgentSpec"
          },
          {
            "properties": {
              "name": {
                "$ref": "#/components/schemas/ResourceName"
              }
            },
            "required": [
              "name"
            ],
            "type": "object"
          }
        ],
        "description": "Complete agent definition used inline on a session or saved as a named agent."
      },
...
      "CreateAgentResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Agent"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "GetCapabilitiesResponse": {
        "properties": {
          "data": {
            "properties": {
              "sandbox": {
                "properties": {
                  "enabled": {
                    "description": "Whether sandboxes can be created: a provider is configured and its sandbox image is synced to the provider.",
                    "type": "boolean"
                  },
                  "reason": {
                    "description": "Present when sandbox is disabled. Explains why.",
                    "type": "string"
                  }
                },
                "required": [
                  "enabled"
                ],
                "type": "object"
              },
              "settings": {
                "properties": {
                  "enabled": {
                    "description": "Whether the settings UI/API is enabled.",
                    "type": "boolean"
                  }
                },
                "required": [
                  "enabled"
                ],
                "type": "object"
              },
              "skill": {
                "properties": {
                  "enabled": {
                    "description": "Whether skills are available. False when sandbox is not enabled (skills require a sandbox).",
                    "type": "boolean"
                  },
                  "reason": {
                    "description": "Present when skills are disabled. Explains why.",
                    "type": "string"
                  }
                },
                "required": [
                  "enabled"
                ],
                "type": "object"
              }
            },
            "required": [
              "sandbox",
              "skill",
              "settings"
            ],
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "GetSessionResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Session"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "ListAgentsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Agent"
            },
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "ListAvailableMcpServersResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/McpServerReadEntry"
            },
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "ListModelsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Model"
            },
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "ListSessionsOrder": {
        "default": "desc",
        "description": "Sort sessions by creation time. Defaults to \"desc\".",
        "enum": [
          "asc",
          "desc"
        ],
        "type": "string"
      },
      "ListSessionsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Session"
            },
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/TokenPagination"
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "type": "object"
      },
...
      "RequestErrorResponse": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "description": "Optional machine-readable error code; null when not applicable.",
                "type": [
                  "string",
                  "null"
                ]
              },
              "message": {
                "description": "Human-readable explanation of the failure.",
                "type": "string"
              },
              "param": {
                "description": "Optional request field that caused the error; null when not field-specific.",
                "type": [
                  "string",
                  "null"
                ]
              },
              "type": {
                "description": "Optional error category (e.g. validation vs conflict).",
                "type": "string"
              }
            },
            "required": [
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "description": "ID token (`Authorization: Bearer <id_token>`). Required on protected routes. Browser sessions may use the HttpOnly `id_token` cookie instead.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HTTP API for the TrueForge agent server (`/api/v1`). Interactive docs are served at `/api/v1/docs` (OpenAPI JSON at `/api/v1/openapi.json`).\n\n**Authentication:** Standalone deployments (no OIDC) accept requests without credentials — middleware stamps a local default user. When OIDC is configured, protected routes require a valid `id_token` cookie or `Authorization: Bearer` ID token. There is no built-in API-key scheme; pass custom headers only if your reverse proxy or IdP layer requires them.\n\nCovers DB-backed sessions, the agent registry, settings catalogs, and model/MCP/skill/sandbox providers.",
    "title": "TrueForge API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/agents": {
      "get": {
        "description": "All configured agents for the tenant.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAgentsResponse"
                }
              }
            },
            "description": "All configured agents."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "OIDC is configured and the request has no valid session cookie."
          }
        },
        "summary": "List agents",
        "tags": [
          "Agents"
        ],
        "x-fern-sdk-group-name": [
          "agents"
        ],
        "x-fern-sdk-method-name": "list"
      },
      "post": {
        "description": "Creates an agent and allocates an immutable id. Fails if `name` is already taken. Name cannot be changed later.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAgentResponse"
                }
              }
            },
            "description": "The created agent."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "Invalid request body or unknown model/MCP/skill refs."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "An agent with this name already exists."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "The agent spec is valid but requires a capability this server does not provide (e.g. sandbox or skills)."
          }
        },
        "summary": "Create an agent",
        "tags": [
          "Agents"
        ],
        "x-fern-sdk-group-name": [
          "agents"
        ],
        "x-fern-sdk-method-name": "create"
      }
    },
...
    "/api/v1/capabilities": {
      "get": {
        "description": "Report optional runtime capabilities available for this tenant.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCapabilitiesResponse"
                }
              }
            },
            "description": "Server capabilities."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "OIDC is configured and the request has no valid session cookie."
          }
        },
        "summary": "Get server capabilities",
        "tags": [
          "Capabilities"
        ],
        "x-fern-sdk-group-name": [
          "server"
        ],
        "x-fern-sdk-method-name": "get_capabilities"
      }
    },
    "/api/v1/mcp-servers": {
      "get": {
        "description": "MCP servers as a slim name/url list for the composer. No auth or auth_status.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAvailableMcpServersResponse"
                }
              }
            },
            "description": "All MCP servers (chat projection)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "OIDC is configured and the request has no valid session cookie."
          }
        },
        "summary": "List MCP servers for chat",
        "tags": [
          "MCP Servers"
        ],
        "x-fern-sdk-group-name": [
          "mcpServers"
        ],
        "x-fern-sdk-method-name": "list"
      }
    },
...
    "/api/v1/models": {
      "get": {
        "description": "Models across all configured model providers, addressed by fully qualified name `name/model_name`.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListModelsResponse"
                }
              }
            },
            "description": "All models of all configured providers."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "OIDC is configured and the request has no valid session cookie."
          }
        },
        "summary": "List models",
        "tags": [
          "Models"
        ],
        "x-fern-sdk-group-name": [
          "models"
        ],
        "x-fern-sdk-method-name": "list"
      }
    },
    "/api/v1/sessions": {
      "get": {
        "description": "List the caller's sessions (newest first by default), token-paginated. Results are scoped to the authenticated identity via the session store's `created_by` filter (not a client query param). Optional `agent_id` filters to sessions bound to that named agent. Pass `page_token` to fetch the next page, keeping the other query params constant.",
        "parameters": [
          {
            "description": "Page size. Defaults to 10, max 100.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "description": "Page size. Defaults to 10, max 100.",
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Sort sessions by creation time. Defaults to \"desc\".",
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ListSessionsOrder"
            }
          },
          {
            "description": "Opaque token from a previous response `next_page_token`.",
            "in": "query",
            "name": "page_token",
            "required": false,
            "schema": {
              "description": "Opaque token from a previous response `next_page_token`.",
              "type": "string"
            }
          },
          {
            "description": "Inclusive lower bound on `created_at` (ISO-8601 / RFC 3339).",
            "in": "query",
            "name": "start_timestamp",
            "required": false,
            "schema": {
              "description": "Inclusive lower bound on `created_at` (ISO-8601 / RFC 3339).",
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Inclusive upper bound on `created_at` (ISO-8601 / RFC 3339).",
            "in": "query",
            "name": "end_timestamp",
            "required": false,
            "schema": {
              "description": "Inclusive upper bound on `created_at` (ISO-8601 / RFC 3339).",
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "When set, only sessions bound to this agent id are returned.",
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "description": "When set, only sessions bound to this agent id are returned.",
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSessionsResponse"
                }
              }
            },
            "description": "Paginated sessions."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "Invalid query parameters or page token."
          }
        },
        "summary": "List sessions",
        "tags": [
          "Sessions"
        ],
        "x-fern-pagination": {
          "cursor": "$request.page_token",
          "next_cursor": "$response.pagination.next_page_token",
          "results": "$response.data"
        },
        "x-fern-sdk-group-name": [
          "sessions"
        ],
        "x-fern-sdk-method-name": "list"
      },
      "post": {
        "description": "Create a session with `agent` as either `{ name }` (named registry binding) or `{ spec: AgentSpec }` (inline). Named sessions snapshot the agent name at create and resolve the live agent on each turn. Responses use `{ type: \"reference\", name, id }` or `{ type: \"inline\", spec }`.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSessionResponse"
                }
              }
            },
            "description": "Session created."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "Invalid request body."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "Named agent not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "The agent spec is valid but references a resource this server does not provide (e.g. model, MCP server, skill, or sandbox)."
          }
        },
        "summary": "Create a session",
        "tags": [
          "Sessions"
        ],
        "x-fern-sdk-group-name": [
          "sessions"
        ],
        "x-fern-sdk-method-name": "create"
      }
    },
...
    "/api/v1/settings/mcp-servers": {
      "get": {
        "description": "All MCP servers with nested auth_status (settings / admin projection).",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListMcpServersResponse"
                }
              }
            },
            "description": "All MCP servers."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "OIDC is configured and the request has no valid session cookie."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "OIDC is configured and the caller is authenticated but not an admin."
          }
        },
        "summary": "List MCP servers",
        "tags": [
          "MCP Servers"
        ],
        "x-fern-sdk-group-name": [
          "settings",
          "mcpServers"
        ],
        "x-fern-sdk-method-name": "list"
      },
      "put": {
        "description": "Full upsert keyed by `name`: creates the server or replaces its manifest. Does not start DCR or modify stored oauth_server / oauth_client columns.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerManifest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PutMcpServerResponse"
                }
              }
            },
            "description": "The saved MCP server with stub auth_status."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "Invalid request body."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "The server cannot satisfy `auth.type: dcr` (e.g. it advertises no registration_endpoint)."
          }
        },
        "summary": "Create or replace an MCP server",
        "tags": [
          "MCP Servers"
        ],
        "x-fern-sdk-group-name": [
          "settings",
          "mcpServers"
        ],
        "x-fern-sdk-method-name": "upsert"
      }
    },
...
    "/api/v1/settings/skills": {
      "get": {
        "description": "All configured skills with full manifests (settings / admin projection).",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListConfiguredSkillsResponse"
                }
              }
            },
            "description": "All configured skills."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "OIDC is configured and the request has no valid session cookie."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "OIDC is configured and the caller is authenticated but not an admin."
          }
        },
        "summary": "List configured skills",
        "tags": [
          "Skills"
        ],
        "x-fern-sdk-group-name": [
          "settings",
          "skills"
        ],
        "x-fern-sdk-method-name": "list"
      },
      "put": {
        "description": "Full upsert keyed by `name`: creates the skill or replaces its entire manifest.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillManifest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PutSkillResponse"
                }
              }
            },
            "description": "The saved skill."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "Invalid request body."
          }
        },
        "summary": "Create or replace a skill",
        "tags": [
          "Skills"
        ],
        "x-fern-sdk-group-name": [
          "settings",
          "skills"
        ],
        "x-fern-sdk-method-name": "upsert"
      }
    },
...
    "/api/v1/skills": {
      "get": {
        "description": "Configured skills as a slim name/description list for the composer.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAvailableSkillsResponse"
                }
              }
            },
            "description": "All configured skills (chat projection)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "OIDC is configured and the request has no valid session cookie."
          }
        },
        "summary": "List skills for chat",
        "tags": [
          "Skills"
        ],
        "x-fern-sdk-group-name": [
          "skills"
        ],
        "x-fern-sdk-method-name": "list"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
}